refactor: changes to structure and code optimization
