+++ package.json
@@ -15,6 +15,7 @@
     "react-router-dom": "^6.15.0"
   },
   "devDependencies": {
+    "@types/node": "^20.6.3",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
