+++ src/components/SideMenu/SideMenu.scss
