+++ src/components/SideMenu/images/BackgroundImages.tsx
@@ -1,0 +1,15 @@
+import WomanLaptop from "../../../assets/images/woman-hold-laptop.svg";
+import WomanLaptopWhite from "../../../assets/images/woman-hold-laptop-white.svg";
+import WomanHappy from "../../../assets/images/woman-happy.svg";
+import BoyTablet from "../../../assets/images/boy-hold-tablet.svg";
+import ManLaptop from "../../../assets/images/man-hold-laptop.svg";
+
+const backgroundImages = {
+  WomanLaptop,
+  WomanLaptopWhite,
+  WomanHappy,
+  BoyTablet,
+  ManLaptop,
+};
+
+export default backgroundImages;
