+++ src/pages/Interest/InterestSuccess.tsx
@@ -1,53 +1,16 @@
-import "../../components/css/Interest.scss";
-import SideMenu from "../../components/SideMenu";
+import "../Interest/Interest.scss";
+import SideMenu from "../../components/SideMenu/SideMenu";
 import VictoryPhoto from "../../assets/images/victory-success.svg";
 import { FormPage } from "../../types/Types";
-
-const PageStates = {
-  Register: "register",
-  Photo: "photo",
-  Interest: "interest",
-  InterestPick: "interestPick",
-  InterestSuccess: "interestSuccess"
-};
-
-const ContinueButton = ({
-  formHomePage
-}: {
-  formHomePage: React.Dispatch<React.SetStateAction<FormPage>>;
-}) => (
-  <button
-    className="continue"
-    onClick={() =>
-      formHomePage({
-        register: true,
-        photo: false,
-        interest_page: false,
-        interest_pick: false,
-        interest_success: false
-      })
-    }
-  >
-    Continue
-  </button>
-);
 
 const InterestSuccess = ({
   formHomePage
 }: {
   formHomePage: React.Dispatch<React.SetStateAction<FormPage>>;
 }) => {
-  const currentPage = PageStates.InterestSuccess;
-
   return (
     <div className="success-width sidemenu-container">
-      <SideMenu
-        registerPage={currentPage === PageStates.Register}
-        photoPage={currentPage === PageStates.Photo}
-        interestPage={currentPage === PageStates.Interest}
-        interestPick={currentPage === PageStates.InterestPick}
-        interestSuccess={currentPage === PageStates.InterestSuccess}
-      />
+      <SideMenu currentPage="interest_success" formHomePage={formHomePage} />
       <div className="success-column">
         <div className="interest-container">
           <div className="interest-column">
@@ -62,7 +25,20 @@ const InterestSuccess = ({
               className="victory-success"
             />
 
-            <ContinueButton formHomePage={formHomePage} />
+            <button
+              className="continue"
+              onClick={() =>
+                formHomePage({
+                  register: true,
+                  photoPage: false,
+                  interest_page: false,
+                  interest_pick: false,
+                  interest_success: false
+                })
+              }
+            >
+              Continue
+            </button>
           </div>
         </div>
       </div>
