+++ src/pages/Photo/Photo.scss
