+++ src/pages/Photo/PhotoPage.tsx
@@ -1,33 +1,18 @@
-import SideMenu from "../components/SideMenu";
-import TopBar from "../components/TopBar";
-import BlankPhoto from "../assets/images/empty-image.svg";
-import "../components/css/Photo.scss";
-import { FormPage } from "../types/Types";
-
-const PageStates = {
-  Register: "register",
-  Photo: "photo",
-  Interest: "interest",
-  InterestPick: "interestPick",
-  InterestSuccess: "interestSuccess"
-};
+import React from "react";
+import SideMenu from "../../components/SideMenu/SideMenu";
+import TopBar from "../../components/TopBar/TopBar";
+import BlankPhoto from "../../assets/images/empty-image.svg";
+import "../Photo/Photo.scss";
+import { FormPage } from "../../types/Types";
 
 const PhotoPage = ({
   formHomePage
 }: {
   formHomePage: React.Dispatch<React.SetStateAction<FormPage>>;
 }) => {
-  const currentPage = PageStates.Photo;
-
   return (
     <div className="sidemenu-container">
-      <SideMenu
-        registerPage={currentPage === PageStates.Register}
-        photoPage={currentPage === PageStates.Photo}
-        interestPage={currentPage === PageStates.Interest}
-        interestPick={currentPage === PageStates.InterestPick}
-        interestSuccess={currentPage === PageStates.InterestSuccess}
-      />
+      <SideMenu currentPage="photoPage" formHomePage={formHomePage} />
       <div>
         <TopBar currentPage="photoPage" formHomePage={formHomePage} /> 
         <div className="photo-container">
@@ -41,7 +26,7 @@ const PhotoPage = ({
               onClick={() =>
                 formHomePage({
                   register: false,
-                  photo: false,
+                  photoPage: false,
                   interest_page: true,
                   interest_pick: false,
                   interest_success: false
@@ -55,7 +40,7 @@ const PhotoPage = ({
               onClick={() =>
                 formHomePage({
                   register: false,
-                  photo: false,
+                  photoPage: false,
                   interest_page: true,
                   interest_pick: false,
                   interest_success: false
