+++ src/types/Types.ts
@@ -1,6 +1,6 @@
 export type FormPage = {
   register: boolean;
-  photo: boolean;
+  photoPage: boolean;
   interest_page: boolean;
   interest_pick: boolean;
   interest_success: boolean;
